refactor: Update workshop controller to return workshop capacity
This commit updates the `WorkshopController` to include a new endpoint `getWorkshopCapacity` that returns the capacity of a workshop. The endpoint is accessed via a GET request to `/capacity/:id`, where `id` is the ID of the workshop. This change improves the functionality of the application by providing a way to retrieve the capacity of a workshop.