feat: Add recoverPassword endpoint to AuthController
This commit adds a new endpoint `recoverPassword` to the `AuthController` class. The endpoint is accessed via a POST request to `/auth/recoverPassword` and allows users to recover their password by sending a recovery token to their email address. This change enhances the functionality of the application by providing a way for users to reset their passwords.